requirements.txt
easydict==1.9
termcolor==1.1.0
ninja
scipy==1.7.1
six==1.16.0
einops==0.3.2
imageio==2.9.0
tqdm==4.62.3
packaging==21.3
av
demjson==1.6
GitPython==3.1.26
iopath==0.1.9
lucid 
matplotlib==3.3.0
moviepy==1.0.3
numpy==1.21.1
omegaconf==2.1.1
opencv_python==4.5.4.60
onnxruntime==1.10.0
Pillow==9.0.0
PyOpenGL==3.1.5
recommonmark==0.7.1
scikit_learn==1.0.2
setuptools==59.5.0
sonnet==0.1.6
torch-fidelity
tensorboard
scikit-image
trimesh
pyrender
pymcubes
pyspng
gdown